Improvements to wp-toolkit plugin
